+++ include/pfnet/constr_AC_FLOW_LIM.h
@@ -3,7 +3,7 @@
  *
  * This file is part of PFNET.
  *
- * Copyright (c) 2015-2017, Tomas Tinoco De Rubira.
+ * Copyright (c) 2015, Tomas Tinoco De Rubira.
  *
  * PFNET is released under the BSD 2-clause license.
  */
@@ -19,11 +19,9 @@
 
 // Function prototypes
 Constr* CONSTR_AC_FLOW_LIM_new(Net* net);
-void CONSTR_AC_FLOW_LIM_init(Constr* c);
 void CONSTR_AC_FLOW_LIM_count_step(Constr* c, Branch* br, int t);
 void CONSTR_AC_FLOW_LIM_analyze_step(Constr* c, Branch* br, int t);
 void CONSTR_AC_FLOW_LIM_eval_step(Constr* c, Branch* br, int t, Vec* v, Vec* ve);
 void CONSTR_AC_FLOW_LIM_store_sens_step(Constr* c, Branch* br, int t, Vec* sA, Vec* sf, Vec* sGu, Vec* sGl);
-void CONSTR_AC_FLOW_LIM_free(Constr* c);
 
 #endif