+++ include/pfnet/constr_BAT_DYN.h
@@ -3,7 +3,7 @@
  *
  * This file is part of PFNET.
  *
- * Copyright (c) 2015-2017, Tomas Tinoco De Rubira.
+ * Copyright (c) 2015, Tomas Tinoco De Rubira.
  *
  * PFNET is released under the BSD 2-clause license.
  */
@@ -16,11 +16,9 @@
 
 // Function prototypes
 Constr* CONSTR_BAT_DYN_new(Net* net);
-void CONSTR_BAT_DYN_init(Constr* c);
 void CONSTR_BAT_DYN_count_step(Constr* c, Branch* br, int t);
 void CONSTR_BAT_DYN_analyze_step(Constr* c, Branch* br, int t);
 void CONSTR_BAT_DYN_eval_step(Constr* c, Branch* br, int t, Vec* v, Vec* ve);
 void CONSTR_BAT_DYN_store_sens_step(Constr* c, Branch* br, int t, Vec* sA, Vec* sf, Vec* sGu, Vec* sGl);
-void CONSTR_BAT_DYN_free(Constr* c);
 
 #endif