+++ src/problem/constr/constr_BAT_DYN.c
@@ -12,20 +12,12 @@
 
 Constr* CONSTR_BAT_DYN_new(Net* net) {
   Constr* c = CONSTR_new(net);
-  CONSTR_set_func_init(c, &CONSTR_BAT_DYN_init);
   CONSTR_set_func_count_step(c, &CONSTR_BAT_DYN_count_step);
   CONSTR_set_func_analyze_step(c, &CONSTR_BAT_DYN_analyze_step);
   CONSTR_set_func_eval_step(c, &CONSTR_BAT_DYN_eval_step);
   CONSTR_set_func_store_sens_step(c, &CONSTR_BAT_DYN_store_sens_step);
-  CONSTR_set_func_free(c, &CONSTR_BAT_DYN_free);
-  CONSTR_init(c);
-  return c;
-}
-
-void CONSTR_BAT_DYN_init(Constr* c) {
-
-  // Init
   CONSTR_set_name(c,"battery dynamics");
+  return c;
 }
 
 void CONSTR_BAT_DYN_count_step(Constr* c, Branch* br, int t) {
@@ -187,8 +179,4 @@ void CONSTR_BAT_DYN_eval_step(Constr* c, Branch* br, int t, Vec* values, Vec* va
 
 void CONSTR_BAT_DYN_store_sens_step(Constr* c, Branch* br, int t, Vec* sA, Vec* sf, Vec* sGu, Vec* sGl) {
   // Nothing for now
-}
-
-void CONSTR_BAT_DYN_free(Constr* c) {
-  // Nothing to do
 }