+++ src/problem/constr/constr_FIX.c
@@ -12,20 +12,12 @@
 
 Constr* CONSTR_FIX_new(Net* net) {
   Constr* c = CONSTR_new(net);
-  CONSTR_set_func_init(c, &CONSTR_FIX_init);
   CONSTR_set_func_count_step(c, &CONSTR_FIX_count_step);
   CONSTR_set_func_analyze_step(c, &CONSTR_FIX_analyze_step);
   CONSTR_set_func_eval_step(c, &CONSTR_FIX_eval_step);
   CONSTR_set_func_store_sens_step(c, &CONSTR_FIX_store_sens_step);
-  CONSTR_set_func_free(c, &CONSTR_FIX_free);
-  CONSTR_init(c);
-  return c;
-}
-
-void CONSTR_FIX_init(Constr* c) {
-
-  // Init
   CONSTR_set_name(c,"variable fixing");
+  return c;
 }
 
 void CONSTR_FIX_count_step(Constr* c, Branch* br, int t) {
@@ -448,8 +440,4 @@ void CONSTR_FIX_eval_step(Constr* c, Branch* br, int t, Vec* values, Vec* values
 
 void CONSTR_FIX_store_sens_step(Constr* c, Branch* br, int t, Vec* sA, Vec* sf, Vec* sGu, Vec* sGl) {
   // Nothing
-}
-
-void CONSTR_FIX_free(Constr* c) {
-  // Nothing to do
 }