+++ src/problem/constr/constr_LOAD_PF.c
@@ -12,20 +12,12 @@
 
 Constr* CONSTR_LOAD_PF_new(Net* net) {
   Constr* c = CONSTR_new(net);
-  CONSTR_set_func_init(c,&CONSTR_LOAD_PF_init);
   CONSTR_set_func_count_step(c,&CONSTR_LOAD_PF_count_step);
   CONSTR_set_func_analyze_step(c,&CONSTR_LOAD_PF_analyze_step);
   CONSTR_set_func_eval_step(c,&CONSTR_LOAD_PF_eval_step);
   CONSTR_set_func_store_sens_step(c,&CONSTR_LOAD_PF_store_sens_step);
-  CONSTR_set_func_free(c,&CONSTR_LOAD_PF_free);
-  CONSTR_init(c);
-  return c;
-}
-
-void CONSTR_LOAD_PF_init(Constr* c) {
-
-  // Init
   CONSTR_set_name(c,"load constant power factor");
+  return c;
 }
 
 void CONSTR_LOAD_PF_count_step(Constr* c, Branch* br, int t) {
@@ -164,8 +156,4 @@ void CONSTR_LOAD_PF_eval_step(Constr* c, Branch* br, int t, Vec* values, Vec* va
 
 void CONSTR_LOAD_PF_store_sens_step(Constr* c, Branch* br, int t, Vec* sA, Vec* sf, Vec* sGu, Vec* sGl) {
   // Nothing for now
-}
-
-void CONSTR_LOAD_PF_free(Constr* c) {
-  // Nothing to do
 }