+++ src/problem/constr/constr_PVPQ_SWITCHING.c
@@ -18,33 +18,14 @@ struct Constr_PVPQ_SWITCHING_Data {
 
 Constr* CONSTR_PVPQ_SWITCHING_new(Net* net) {
   Constr* c = CONSTR_new(net);
-  CONSTR_set_func_init(c, &CONSTR_PVPQ_SWITCHING_init);
   CONSTR_set_func_count_step(c, &CONSTR_PVPQ_SWITCHING_count_step);
   CONSTR_set_func_allocate(c, &CONSTR_PVPQ_SWITCHING_allocate);
-  CONSTR_set_func_clear(c, &CONSTR_PVPQ_SWITCHING_clear);
   CONSTR_set_func_analyze_step(c, &CONSTR_PVPQ_SWITCHING_analyze_step);
   CONSTR_set_func_eval_step(c, &CONSTR_PVPQ_SWITCHING_eval_step);
   CONSTR_set_func_store_sens_step(c, &CONSTR_PVPQ_SWITCHING_store_sens_step);
   CONSTR_set_func_free(c, &CONSTR_PVPQ_SWITCHING_free);
-  CONSTR_init(c);
-  return c;
-}
-
-void CONSTR_PVPQ_SWITCHING_init(Constr* c) {
-
-  // Init
   CONSTR_set_name(c,"PVPQ switching");
-  CONSTR_set_data(c,NULL);
-}
-
-void CONSTR_PVPQ_SWITCHING_clear(Constr* c) {
-
-  // Counters
-  CONSTR_set_A_nnz(c,0);
-  CONSTR_set_A_row(c,0);
-
-  // Flags
-  CONSTR_clear_bus_counted(c);
+  return c;
 }
 
 void CONSTR_PVPQ_SWITCHING_count_step(Constr* c, Branch* br, int t) {
@@ -113,36 +94,15 @@ void CONSTR_PVPQ_SWITCHING_count_step(Constr* c, Branch* br, int t) {
 void CONSTR_PVPQ_SWITCHING_allocate(Constr* c) {
 
   // Local variables
-  int num_constr;
+  Constr_PVPQ_SWITCHING_Data* data;
   int num_vars;
-  int A_nnz;
   Bus* bus;
   Net* net;
   int i;
   int t;
-  Constr_PVPQ_SWITCHING_Data* data;
 
   net = CONSTR_get_network(c);
   num_vars = NET_get_num_vars(net);
-  num_constr = CONSTR_get_A_row(c);
-  A_nnz = CONSTR_get_A_nnz(c);
-
-  // J f
-  CONSTR_set_J(c,MAT_new(0,num_vars,0));
-  CONSTR_set_f(c,VEC_new(0));
-
-  // G u l
-  CONSTR_set_G(c,MAT_new(0,num_vars,0));
-  CONSTR_set_u(c,VEC_new(0));
-  CONSTR_set_l(c,VEC_new(0));
-
-  // b
-  CONSTR_set_b(c,VEC_new(num_constr));
-
-  // A
-  CONSTR_set_A(c,MAT_new(num_constr, // size1 (rows)
-			 num_vars,   // size2 (rows)
-			 A_nnz)); // nnz
 
   // Data (var-dependent)
   CONSTR_PVPQ_SWITCHING_free(c);